Added fractional scaling.

+++ source/window.h
@@ -30,9 +30,9 @@ void WindowDrawLine(Window *window, uint32_t color, uint32_t startX, uint32_t st
 
 void WindowDrawRectangle(Window *window, uint32_t color, uint32_t x, uint32_t y, uint32_t width, uint32_t height);
 
-void WindowDrawSprite(Window *window, Sprite *sprite, uint32_t x, uint32_t y, uint32_t scale);
+void WindowDrawSprite(Window *window, Sprite *sprite, uint32_t x, uint32_t y, float scale);
 
-void WindowDrawText(Window *window, Sprite *font, const char *text, uint32_t color, uint32_t x, uint32_t y, uint32_t scale);
+void WindowDrawText(Window *window, Sprite *font, const char *text, uint32_t color, uint32_t x, uint32_t y, float scale);
 
 uint32_t WindowGetWidth(Window *window);
